requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Pillow==9.5.0
transformers==4.38.1
torch>=2.0.0
TTS==0.17.3
ctransformers==0.2.27
itsdangerous==2.1.2
